requirements.txt
fastapi
uvicorn
loguru
pydantic
pydantic-settings
langchain-google-genai
langchain-core
langchain
langgraph
langgraph-checkpoint-sqlite
pipecat-ai[silero,whisper,elevenlabs]
aiortc
opencv-python
hsemotion
ultralytics
timm<1.0